Extract AudioFieldset
